Removed redundant static_cast<uintptr_t>
at least it seems redundant - if it will be the case - I need to revisit all code everywhere (hooray)

+++ skiko/src/jvmMain/cpp/common/Font.cc
@@ -222,7 +222,7 @@ extern "C" JNIEXPORT jshortArray JNICALL Java_org_jetbrains_skia_FontKt__1nGetSt
 
 extern "C" JNIEXPORT void JNICALL Java_org_jetbrains_skia_FontKt__1nGetUTF32Glyphs
   (JNIEnv* env, jclass jclass, jlong ptr, jintArray uniArr, jint uniCount, jshortArray glyphsActual) {
-    SkFont* instance = reinterpret_cast<SkFont*>(static_cast<uintptr_t>(ptr));
+    SkFont* instance = reinterpret_cast<SkFont*>(ptr);
     std::vector<jshort> glyphs(uniCount);
     jint* uni = env->GetIntArrayElements(uniArr, nullptr);
     instance->unicharsToGlyphs(reinterpret_cast<SkUnichar*>(uni), uniCount, reinterpret_cast<SkGlyphID*>(glyphs.data()));
